fix(scenario-outlines): capture scenarioStepsToRun in closure so that each example run executes the correct function
The "multiple examples" aspect of the scenario outline feature is broken. The synchronous loop over each of the examples causes the test runner to execute the test suite N times, but it does so on each iteration with the _last_ item of example data, rather than the _i_th item.

This change simply creates a closure around the current iteration's step functions, so that they aren't clobbered on the next go-around of the loop.